Fix @jsonflatten in arcadia SDK not working issue
